+++ src/Snippet.php
@@ -3,7 +3,7 @@ namespace ddStringTools;
 
 class Snippet extends \DDTools\Snippet {
 	protected
-		$version = '2.2.0',
+		$version = '2.3.0',
 		
 		$params = [
 			// Defaults